requirements.txt
Flask==2.2.5
requests==2.31.0
python-dotenv==1.0.0
openai==0.27.8
Flask-SQLAlchemy==3.0.3
python-dateutil==2.8.2
Flask-WTF==1.1.1
Werkzeug==2.3.6
aiohttp==3.8.4
emoji==2.6.0
Flask-Login==0.6.2
cryptography==41.0.1
Pillow==9.5.0
PyJWT==2.7.0
gunicorn
aiohttp==3.8.4
cryptography==41.0.1
emoji==2.6.0
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gunicorn
openai==0.27.8
Pillow==9.5.0
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.6
flask
python-dotenv
requests
flask
python-dotenv
requests
aiohttp
cryptography
emoji
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
openai
pillow
pyjwt
python-dateutil
python-dotenv
requests
werkzeug
